FL-1832: install gcc wrappers for other multilib ABIs
Install gcc wrappers and gcc-config files to support using host gcc as
an alternate ABI compiler. For example, gcc on multilib amd64 can be
used as i686-pc-linux-gnu compiler. This makes it possible to use
distcc on an amd64 host for x86 clients. It will also make multilib
cleaner in the future.